change needed requests to post

+++ tobuy/views.py
@@ -9,7 +9,7 @@ User = get_user_model()
 
 @login_required()
 def index(request):
-    qs = MustBuy.objects.filter(user=request.user)
+    qs = MustBuy.objects.filter(user=request.user).order_by('-create_date')
     return render(request, 'tobuy/tobuy.html', {
         'carts': qs,
     })
@@ -50,16 +50,18 @@ def edit_mustbuy(request, id):
     })
 
 @login_required()
-def set_done_mustbuy(request, id):
-    mustbuy = get_object_or_404(MustBuy, id=id)
-    mustbuy.status = True
-    mustbuy.save()
+def change_status_mustbuy(request, id):
+    if request.method == "POST":
+        mustbuy = get_object_or_404(MustBuy, id=id)
+        mustbuy.status = not mustbuy.status
+        mustbuy.save()
 
-    return redirect(reverse("tobuy"))
+        return redirect(reverse("tobuy"))
 
 @login_required()
 def delete_mustbuy(request, id):
-    mustbuy = get_object_or_404(MustBuy, id=id)
-    mustbuy.delete()
+    if request.method == "POST":
+        mustbuy = get_object_or_404(MustBuy, id=id)
+        mustbuy.delete()
 
-    return redirect(reverse("tobuy"))
+        return redirect(reverse("tobuy"))
